39.1.JS to Auto-Update the Select Options{Adding the Options Endpoint}
Thanks to these event listeners, no matter what data we start with - or what data we submit - for the location field, the specificLocationName field choices will update so that everything saves.
The last step is to add some JavaScript! When the form loaded, the location was set to "Near a star". When I change it to "The Solar System", we need to make an Ajax call that will fetch the list of planets and update the option elements.

Adding the Options Endpoint
In ArticleAdminController, let's add a new endpoint for this: public function getSpecificLocationSelect(). Add Symfony's Request object as an argument. Here's the idea: our JavaScript will send the location that was just selected to this endpoint and it will return the new HTML needed for the entire specificLocationName field. So, this won't be a pure API endpoint that returns JSON. We could do that, but because the form is already rendering our HTML, returning HTML simplifies things a bit.
Above the method add the normal @Route() with /admin/article/location-select. And give it a name="admin_article_location_select".
Inside, the logic is kinda cool: create a new Article: $article = new Article(). Next, we need to set the new location onto that. When we make the AJAX request, we're going to add a ?location= query parameter. Read that here with $request->query->get('location').
But, let's back up: we're not creating this Article object so we can save it, or anything like that. We're going to build a temporary form using this Article's data, and render part of it as our response. Check it out: $form = ->createForm(ArticleFormType::class, $article). We know that, thanks to our event listeners - specifically our PRE_SET_DATA event listener - this form will now have the correct specificNameLocation options based on whatever location was just sent to us.
Or, the field may have been removed! Check for that first: if (! $form->has('specificLocationName') then just return new Response() - the one from HttpFoundation - with no content. I'll set the status code to 204, which is a fancy way of saying that the call was successful, but we have no content to send back.
If we do have that field, we want to render it! Return and render a new template: article_admin/_specific_location_name.html.twig. Pass this the form like normal 'articleForm' => $form->createView(). Then, I'll put my cursor on the template name and press alt+enter to make PhpStorm create that template for me. Inside, just say: {{ form_row(articleForm.specificLocationName) }} and that's it.
Yep, we're literally returning just the form row markup for this one field. It's a weird way to use a form, but it works!
Let's go try this out! Copy the new URL, open a new tab and go to http://localhost:8000/admin/article/location-select?location=star
Cool! A drop-down of stars! Try solar_system and... that works too. Excellent